requirements.txt
numpy
seaborn
geopandas
Image
opencv-python
instabot
urllib3
descartes
glob2
shutil
python-firebase
bs4
